+++ core/src/main/java/arez/annotations/CascadeDispose.java
@@ -15,13 +15,13 @@ import java.lang.annotation.Target;
  *
  * <p>If the element annotated is a method then the method must comply with the additional constraints:</p>
  * <ul>
- * <li>Must not be annotated with any other arez annotation</li>
+ * <li>Must not be annotated with any other arez annotation other than {@link Reference} or {@link Observable}</li>
  * <li>Must be final</li>
  * <li>Must have 0 parameters</li>
  * <li>The type of the field must implement {@link arez.Disposable} or must be annotated by {@link ArezComponent}</li>
  * <li>Must not be private</li>
  * <li>Must not be static</li>
- * <li>Must not be abstract</li>
+ * <li>Must not be abstract unless the method is annotated with {@link Reference} or {@link Observable} in which case it MUST be abstract</li>
  * <li>Must not throw exceptions</li>
  * <li>The method must be accessible to the component subclass which means it must not be package access unless it is in the same package as the arez component.</li>
  * </ul>