+++ processor/src/test/resources/expected/com/example/reference/Arez_CascadeDisposeReferenceModel.java
@@ -1,0 +1,159 @@
+package com.example.reference;
+
+import arez.Arez;
+import arez.ArezContext;
+import arez.Component;
+import arez.Disposable;
+import arez.Locator;
+import arez.SafeProcedure;
+import arez.component.DisposeNotifier;
+import arez.component.Identifiable;
+import arez.component.Verifiable;
+import arez.component.internal.ComponentKernel;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import org.realityforge.braincheck.Guards;
+
+@Generated("arez.processor.ArezProcessor")
+final class Arez_CascadeDisposeReferenceModel extends CascadeDisposeReferenceModel implements Disposable, Identifiable<Integer>, Verifiable, DisposeNotifier {
+  private static volatile int $$arezi$$_nextId;
+
+  private final ComponentKernel $$arezi$$_kernel;
+
+  @Nullable
+  private CascadeDisposeReferenceModel.MyEntity $$arezr$$_myEntity;
+
+  Arez_CascadeDisposeReferenceModel() {
+    super();
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> Arez.areReferencesEnabled(), () -> "Attempted to create instance of component of type 'CascadeDisposeReferenceModel' that contains references but Arez.areReferencesEnabled() returns false. References need to be enabled to use this component" );
+    }
+    final ArezContext $$arezv$$_context = Arez.context();
+    final int $$arezv$$_id = ( Arez.areNamesEnabled() || Arez.areRegistriesEnabled() || Arez.areNativeComponentsEnabled() ) ? ++$$arezi$$_nextId : 0;
+    final String $$arezv$$_name = Arez.areNamesEnabled() ? "CascadeDisposeReferenceModel." + $$arezv$$_id : null;
+    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "CascadeDisposeReferenceModel", $$arezv$$_id, $$arezv$$_name, () -> $$arezi$$_nativeComponentPreDispose() ) : null;
+    this.$$arezi$$_kernel = new ComponentKernel( Arez.areZonesEnabled() ? $$arezv$$_context : null, Arez.areNamesEnabled() ? $$arezv$$_name : null, $$arezv$$_id, Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNativeComponentsEnabled() ? null : this::$$arezi$$_preDispose, null, null, true, false, false );
+    this.$$arezi$$_kernel.componentConstructed();
+    this.$$arezi$$_link_myEntity();
+    this.$$arezi$$_kernel.componentReady();
+  }
+
+  @Nonnull
+  final Locator $$arezi$$_locator() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named '$$arezi$$_locator' invoked on uninitialized component of type 'CascadeDisposeReferenceModel'" );
+    }
+    return this.$$arezi$$_kernel.getContext().locator();
+  }
+
+  final int $$arezi$$_id() {
+    return this.$$arezi$$_kernel.getId();
+  }
+
+  @Override
+  @Nonnull
+  public final Integer getArezId() {
+    return $$arezi$$_id();
+  }
+
+  private void $$arezi$$_preDispose() {
+    Disposable.dispose( $$arezr$$_myEntity );
+  }
+
+  private void $$arezi$$_nativeComponentPreDispose() {
+    this.$$arezi$$_preDispose();
+    this.$$arezi$$_kernel.notifyOnDisposeListeners();
+  }
+
+  @Override
+  public void addOnDisposeListener(@Nonnull final Object key, @Nonnull final SafeProcedure action) {
+    this.$$arezi$$_kernel.addOnDisposeListener( key, action );
+  }
+
+  @Override
+  public void removeOnDisposeListener(@Nonnull final Object key) {
+    this.$$arezi$$_kernel.removeOnDisposeListener( key );
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return this.$$arezi$$_kernel.isDisposed();
+  }
+
+  @Override
+  public void dispose() {
+    this.$$arezi$$_kernel.dispose();
+  }
+
+  @Override
+  public void verify() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'verify' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? '?' : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() && Arez.isVerifyEnabled() ) {
+      Guards.apiInvariant( () -> this == $$arezi$$_locator().findById( CascadeDisposeReferenceModel.class, $$arezi$$_id() ), () -> "Attempted to lookup self in Locator with type CascadeDisposeReferenceModel and id '" + $$arezi$$_id() + "' but unable to locate self. Actual value: " + $$arezi$$_locator().findById( CascadeDisposeReferenceModel.class, $$arezi$$_id() ) );
+      final int $$arezv$$_myEntityId = this.getMyEntityId();
+      final CascadeDisposeReferenceModel.MyEntity $$arezv$$_myEntity = this.$$arezi$$_locator().findById( CascadeDisposeReferenceModel.MyEntity.class, $$arezv$$_myEntityId );
+      Guards.apiInvariant( () -> null != $$arezv$$_myEntity, () -> "Reference named 'myEntity' on component named '" + this.$$arezi$$_kernel.getName() + "' is unable to resolve entity of type com.example.reference.CascadeDisposeReferenceModel.MyEntity and id = " + getMyEntityId() );
+    }
+  }
+
+  @Override
+  CascadeDisposeReferenceModel.MyEntity getMyEntity() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'getMyEntity' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? '?' : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != $$arezr$$_myEntity, () -> "Nonnull reference method named 'getMyEntity' invoked on component named '" + this.$$arezi$$_kernel.getName() + "' but reference has not been resolved yet is not lazy. Id = " + getMyEntityId() );
+    }
+    return this.$$arezr$$_myEntity;
+  }
+
+  private void $$arezi$$_link_myEntity() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named '$$arezi$$_link_myEntity' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? '?' : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    final int id = this.getMyEntityId();
+    this.$$arezr$$_myEntity = this.$$arezi$$_locator().findById( CascadeDisposeReferenceModel.MyEntity.class, id );
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != $$arezr$$_myEntity, () -> "Reference named 'myEntity' on component named '" + this.$$arezi$$_kernel.getName() + "' is unable to resolve entity of type com.example.reference.CascadeDisposeReferenceModel.MyEntity and id = " + getMyEntityId() );
+    }
+  }
+
+  private void $$arezi$$_delink_myEntity() {
+    this.$$arezr$$_myEntity = null;
+  }
+
+  @Override
+  public final int hashCode() {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      return Integer.hashCode( $$arezi$$_id() );
+    } else {
+      return super.hashCode();
+    }
+  }
+
+  @Override
+  public final boolean equals(final Object o) {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      if ( o instanceof Arez_CascadeDisposeReferenceModel ) {
+        final Arez_CascadeDisposeReferenceModel that = (Arez_CascadeDisposeReferenceModel) o;
+        return $$arezi$$_id() == that.$$arezi$$_id();
+      } else {
+        return false;
+      }
+    } else {
+      return super.equals( o );
+    }
+  }
+
+  @Override
+  public final String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + this.$$arezi$$_kernel.getName() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}