+++ processor/src/test/resources/input/com/example/reference/CascadeDisposeReferenceModel.java
@@ -1,0 +1,25 @@
+package com.example.reference;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+import arez.annotations.Reference;
+import arez.annotations.ReferenceId;
+
+@ArezComponent
+abstract class CascadeDisposeReferenceModel
+{
+  @Reference
+  @CascadeDispose
+  abstract MyEntity getMyEntity();
+
+  @ReferenceId
+  int getMyEntityId()
+  {
+    return 0;
+  }
+
+  @ArezComponent( allowEmpty = true )
+  static abstract class MyEntity
+  {
+  }
+}